Little bits of improvements and refactoring
