+++ components/app-shell.js
@@ -1,17 +1,17 @@
 import Link from 'next/link';
 import Head from 'next/head';
-import styles from './_styles/app-shell.module.scss';
+import styles from '../styles/app-shell.module.scss';
 
 const AppShell = (props)=> {
   return <>
     <Head key={'Site-wide <head> content'}>
-      <title>Fẹranmi Akinlade | Front-end Engineer</title>
+      <title>Feranmi Akinlade | Front-end Engineer</title>
     </Head>
 
     <header key={'Titlebar'} className={styles['site-header']}>
       <h1 className={styles['page-title']}>
         <Link href={{ pathname: '/', hash: '#about' }}>
-          <a>Fẹranmi Akinlade</a>
+          <a>Feranmi Akinlade</a>
         </Link>
       </h1>
 
@@ -50,12 +50,12 @@ const AppShell = (props)=> {
             target="_blank" rel="noopener noreferrer">
           GitHub
         </a> &nbsp;
-        <a href="https://stackoverflow.com/users/story/7665742"
+        {/* <a href="https://stackoverflow.com/users/story/7665742"
             target="_blank" rel="noopener noreferrer"
             title="My developer story on Stack Overflow">
           Stack Overflow
-        </a> &nbsp;
-        <a href="https://www.linkedin.com/in/akinlade-feranmi-0b6bbb170"
+        </a> &nbsp; */}
+        <a href="https://www.linkedin.com/in/uxferanmi/"
             target="_blank" rel="noopener noreferrer" title="My LinkedIn profile">
           LinkedIn
         </a> &nbsp;
