+++ pages/_app.js
@@ -1,6 +1,6 @@
 // import App from 'next/app'
-import './_styles/global.scss';
-import AppShell from './_app-shell';
+import '../styles/global.scss';
+import AppShell from '../components/app-shell';
 
 function MyApp({ Component, pageProps }) {
 	return <AppShell>
