+++ styles/app-shell.module.scss
@@ -7,7 +7,7 @@
   align-items: center;
   justify-content: space-between;
   flex-wrap: wrap;
-  padding: 10px 30px;
+  padding: 10px 10vw;
   // margin-bottom: 20px;
   background-color: white;
 
@@ -30,7 +30,7 @@
 
   &__item {
     list-style: none;
-    margin-left: 20px;
+    margin-right: 20px;
     white-space: nowrap;
 
     a, a:visited {
