+++ styles/home.module.scss
@@ -1,6 +1,6 @@
 .section {
   padding-top: 120px;
-  &:not(:last-of-type) {
+  &:not(:last-child) {
     margin-bottom: -60px;
   }
   display: flex;
@@ -12,9 +12,11 @@
 
   h2 {
     margin-bottom: 15px;
+    max-width: 600px;
   }
 
   p {
     margin-bottom: 20px;
+    max-width: 600px;
   }
 }
